+++ src/helper.c
@@ -4,7 +4,7 @@ void	charPrintHelper(char c){
   write(1, &c, 1);
 }
 
-size_t stringLengthHelper(const string s){
+long int stringLengthHelper(const string s){
   string	bfs;
 
   for (bfs = s; *bfs != '\0'; ++bfs);
